+++ src/backend/src/services/onboarding.services.ts
@@ -238,19 +238,4 @@ export default class OnboardingServices {
       data: { dateDeleted: new Date(), userDeletedId: deleter.userId }
     });
   }
-
-  static async setOnboardingText(submitter: User, text: string, organization: Organization) {
-    if (!(await userHasPermission(submitter.userId, organization.organizationId, isAdmin))) {
-      throw new AccessDeniedAdminOnlyException('non-admin tried to update onboarding text');
-    }
-
-    await prisma.organization.update({
-      where: {
-        organizationId: organization.organizationId
-      },
-      data: {
-        onboardingText: text
-      }
-    });
-  }
 }
